feat: autotest pls

+++ controllers/cards.js
@@ -27,10 +27,12 @@ const createCard = async (req, res, next) => {
 const deleteCard = async (req, res, next) => {
   try {
     const card = await Card.findById(req.params.id);
-    if (card.owner.toString() === req.user._id) {
-      const deletedcard = await Card.findByIdAndDelete(req.params.id);
-      if (deletedcard) {
-        return res.send({ data: deletedcard });
+    if (card) {
+      if (card.owner.toString() === req.user._id) {
+        const deletedcard = await Card.findByIdAndDelete(req.params.id);
+        if (deletedcard) {
+          return res.send({ data: deletedcard });
+        }
       }
     } else {
       return next(new ForbiddenError('Нет прав на удаление'));
@@ -60,7 +62,7 @@ const likeCard = async (req, res, next) => {
     return next(new NotFoundError('Карточка не найдена'));
   } catch (err) {
     if (err.name === 'CastError') {
-      return next(new NotFoundError('Карточка не найдена'));
+      return next(new BadRequestError('Карточка не найдена'));
     }
     return next(err);
   }
@@ -82,7 +84,7 @@ const dislikeCard = async (req, res, next) => {
     return next(new NotFoundError('Карточка не найдена'));
   } catch (err) {
     if (err.name === 'CastError') {
-      return next(new NotFoundError('Карточка не найдена'));
+      return next(new BadRequestError('Карточка не найдена'));
     }
     return next(err);
   }
